backend/requirements.txt
fastapi>=0.115.0,<1
uvicorn[standard]>=0.32.0,<1
sqlalchemy>=2.0.36,<3

# Data pipeline deps (Step 2)
numpy>=2.1.0,<3
beautifulsoup4>=4.12.3,<5
pypdf>=5.1.0,<6
google-play-scraper>=1.2.7,<2
google-api-python-client>=2.150.0,<3
google-auth>=2.35.0,<3

# Dev / test
httpx>=0.27.0,<1
pytest>=8.3.0,<9
